@@ -85,19 +84,20 @@
     "source-map-loader": "^3.0.0",
     "style-loader": "^3.3.1",
     "tailwindcss": "^3.0.2",
-    "terser-webpack-plugin": "^5.2.5",
-    "ts-jest": "^29.1.1",
-    "typescript": "^5.7.3",
-    "uuid": "^11.0.5",
+    "terser-webpack-plugin": "^5.3.14",
+    "ts-jest": "^29.2.6",
+    "typescript": "^5.8.2",
+    "uuid": "^11.1.0",
     "web-vitals": "^2.1.4",
-    "webpack": "^5.97.1",
+    "webpack": "^5.98.0",
     "webpack-dev-server": "^5.2.0",
-    "webpack-manifest-plugin": "^5.0.0",
+    "webpack-manifest-plugin": "^5.0.1",
     "workbox-webpack-plugin": "^7.3.0"
   },
   "scripts": {
     "start": "node scripts/start.js",
     "build": "node scripts/build.js",
+    "test:m": "jest",
     "test": "node scripts/test.js --verbose",
     "test:watch": "node scripts/test.js --verbose --watch",
     "docs": "jsdoc -c jsdoc.conf.json -R README.md -r",
