@@ -1,12 +1,13 @@
-'use strict';
+// 'use strict';
 
-const path = require('path');
-const camelcase = require('camelcase');
+import path from 'path';
+// const camelcase = require('camelcase');
+import camelcase from 'camelcase';
 
 // This is a custom Jest transformer turning file imports into filenames.
 // http://facebook.github.io/jest/docs/en/webpack.html
 
-module.exports = {
+export default {
   process(src, filename) {
     const assetFilename = JSON.stringify(path.basename(filename));
 
