@@ -5,9 +5,7 @@
 
 module.exports = {
   process() {
-    return {
-      code: 'module.exports = {};'
-    };
+    return 'module.exports = {};';
   },
   getCacheKey() {
     // The output is always the same.
